Refactor octree server to export grpc server
